+++ main/src/main/resources/org/clulab/numeric/HOLIDAY.tsv
@@ -1,4 +1,14 @@
-# list of holidays
+#
+# List of holidays in the United States extracted from the Jollyday api.
+# This list can be obtained with the following snippet:
+#
+#   import de.jollyday.{HolidayManager, HolidayCalendar}
+#   import collection.JavaConverters._
+#   val holidayManager = HolidayManager.getInstance(HolidayCalendar.UNITED_STATES)
+#   holidayManager.getCalendarHierarchy.getChildren.keySet.asScala.
+#                   map(holidayManager.getHolidays(2021, _).asScala).flatten.map(_.getDescription)
+#
+#
 Lincoln's Birthday
 Easter Monday
 Washington's Birthday
