+++ main/src/main/resources/org/clulab/numeric/dates.yml
@@ -138,5 +138,5 @@ rules:
     example: "It was Christmas 2021."
     action: mkDateMentionHoliday
     pattern: |
-      @holiday:PossibleHoliday @year:PossibleYear?
+      @holiday:Holiday @year:PossibleYear?
